apply suggestions from code review
- refactor for consistency
- beginStopSequence: rework rule logic to no longer maintain a list on integer
- rework tests for clarity
- additional unit test
- improve documentation

TODO: finish test refactor + additional unit test for non time points fast and non fast travel
